Remove missnamed \'Labrynth\' file, replace maze with newMaze and rename to maze
